Fixes Keras2 DeepLap Tests

+++ keras_cv/layers/spatial_pyramid.py
@@ -165,7 +165,9 @@ class SpatialPyramidPooling(keras.layers.Layer):
             result.append(temp)
 
         result[-1] = keras.layers.Resizing(
-            inputs.shape[1], inputs.shape[2], interpolation="bilinear"
+            ops.cast(inputs.shape[1], dtype="int32"),
+            ops.cast(inputs.shape[2], dtype="int32"),
+            interpolation="bilinear",
         )(result[-1])
 
         result = ops.concatenate(result, axis=-1)
